map检查障碍物的接口改成float

+++ yhge/Isometric/tile/ISOTileMap.cpp
@@ -279,7 +279,7 @@ void ISOTileMap::showCoordLine()
 /**
  *检查某个单元是否可以通过
  */
-bool ISOTileMap::isWorkable(int x,int y)
+bool ISOTileMap::isWorkable(float x,float y)
 {
     //根据具体地图类型来实现。比如用某个层来标识路径情况。
     //如果地图上可移动物体是不能穿透，则也要算在路径里。
